I hope I break nothing important. Edit(MR): You did.

See merge request mreinsch1/pm_dungeon!24

+++ src/items/weapons/Weapon.java
@@ -2,11 +2,24 @@ package items.weapons;
 
 import items.EquipableItem;
 import main.Game;
-
+/**
+ * Weapon base class.
+ * <p>
+ *   Contains everything that describes a shield.
+ * </p>
+ */
 public abstract class Weapon extends EquipableItem {
   protected float attackDamageModifier;
+  /**
+   *  Returns the attackDamageModifier value of the shield which can be used for display purposes
+   *  @return attackDamageModifier of the shield
+   */
   public float getAttackDamageModifier(){return attackDamageModifier;};
   protected float hitChanceModifier;
+  /**
+   *  Returns the hitChanceModifier value of the shield which can be used for display purposes
+   *  @return hitChanceModifier of the shield
+   */
   public float getHitChanceModifier(){return hitChanceModifier;};
   protected int condition;
   public boolean reduceCondition(int reduce) {
@@ -16,8 +29,17 @@ public abstract class Weapon extends EquipableItem {
     }
     return true;
   }
-  public int getCondition(){return condition;};
-
+  /**
+   *  Returns the condition value of the shield which can be used for display purposes
+   *  @return condition of the shield
+   */
+  public float getCondition(){return condition;};
+  /**
+   * Constructor of the Weapon base class.
+   * <p>
+   * This constructor will instantiate the animations and read all required texture data.
+   * </p>
+   */
   public Weapon(Game game) {
     super(game);
   }
